Fix HOME button and menu navigation crashes
- Replace corrupted local function pointers with NULL to prevent crashes
- Add null safety checks for menu init_function and action_function calls
- Implement inline handling for critical menu actions (Load/Save State, Return to Game, Quit)
- Fix format specifier warning in main.c sprintf call
- Add debug logging wrapper for HOME button menu transitions

The root cause was local auto functions being used as function pointers in menu structures.
When PSP module reloads between game/menu transitions, these addresses become invalid,
causing BadVAddr 6C666408 crashes when menu tries to call corrupted function pointers.

+++ source/src/main.c
@@ -550,7 +550,7 @@ static void synchronize(void)
     if (psp_fps_debug != 0)
     {
       char print_buffer[16];
-      sprintf(print_buffer, "%02d(%02d)", fps, frames_drawn);
+      sprintf(print_buffer, "%02ld(%02ld)", (long)fps, (long)frames_drawn);
 //    sprintf(print_buffer, "%02d(%02d)", fps % 100, frames_drawn);
       print_string(print_buffer, 0, 0, COLOR15_WHITE, COLOR15_BLACK);
     }
